requirements.txt
google-generativeai
requests
beautifulsoup4
mlflow
dvc
tqdm
python-dotenv
